requirements.txt
PyNaCl
aiohttp==3.7.4.post0
astroid==2.5.1
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
colorama==0.4.4
discord.py==1.7.3
idna==3.2
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
multidict==5.1.0
pylint==2.7.2
toml==0.10.2
typing-extensions==3.10.0.0
wrapt==1.12.1
yarl==1.6.3
requests==2.15.1
discord_components==1.1.4
DiscordUtils==1.3.4
youtube_dl==2021.6.6
dblpy==0.4.0
giphy_client==1.0.0
discord.py[voice]
pistonapi==1.0.1
discord-components==1.1.4

